Grey out & disallow unusable powers.
Greys out powers on action bar that cannot be used due to:
* required weapon not euipped
* lacking required discipline -- such powers can be put in the actionbar
  when using an item that increases an attribute (e.g., Scathelocke's
  Spellbook), and the item is later removed, dropping the discipline
  below the required value.
* character being stunned
* character being dead
* character not having cleaned their room or done their homework

Unfortunately, there are a number of places where similar checks are
made. I've added a few functions to help centralize this logic, but I
haven't implemented it well.  Example, calling StatBlock::canUsePower()
in Avatar::logic(), but I haven't removed subsequent code that makes
some of the same checks.