Refactor: Uncopyable
class Uncopyable was added in the same patch that added the Vec2
template class.  Uncopyable is a simple mechanism to cause a class to
become uncopyable, eliminating the risk of un-intended copy or
assignment using the compiler-generated copy constructor or assignment
operator.  Generally, any class that has a pointer data member should be
Uncopyable (there are exceptions of course).  While doing other
refactoring, I actually ended up with a problem because I accidentally
copied the map object when I meant to store a reference to it.
Employing Uncopyable solves this problem