Refactor: CombatText
This patch transforms struct CombatTextItem into a full class with
private members and a public constructor (although it is friends with
class CombatText).  This greatly cleans up sources for the construction
of new CombatTextItem objects.

Double creation of CombatTextItems has been factored out and the
CombatText's vector replaced with a list.  Finally, the entire display
mechanism is changed to handle messages with varying display times,
which may be desirable at some point in the future.