Tweak: Knights interrupted less in melee combat.
This patch affects game play.  When a PC is hit (Avatar::takeHit()) and
are engaged in the AVATAR_MELEE stance (action/skill) while wielding a
shield, they are granted a chance to not be interrupted.  The chance of
avoiding interruption is 20% plus an additional 20% per point of
absorption with a maxiumum chance of 90%.  When a shield as an
absorption range, the average value is used.