Refactor: GameStates and MenuManager
Pointers to gui objects that are constructed via a call to new by the
object and destroyed with a delete in the destructor have been
refactored.  Now, these classes will simply use composition and allow
these UI components to be naturally created and destroyed as data
members.  This is far cleaner, eliminates the possibility of memory
leaks and other pointer and memory errors and is much less code.

I have changed visibility of some members, making the private and adding
getters & setters as needed, which improves encapsulation.