Refactor: Items
This patch is similar to the Powers refactoring.  An int id data member
is added to the Item class and Items are now referenced by a const Item
*pointer rather than an item id.  This simplifies many things -- for
one, when you need item data, you don't have to refer to the ItemManager
anymore.  When "no item" needs to be specified, the const Item *pointer
is simply set to NULL and this condition is now checked for.

ItemStorage's operator[] is now inline and exists in two forms: a const
and non-const, to reduce the need for a non-const ItemManager reference.

This patch intruduces a bug that still needs to be fixed: somewhere,
picking up loot (off of the floor) can get screwed up and you pick up,
picking up the wrong item and even duplicating items -- oops.