Use compile-time constants for constant settings
Many of the values in Settings.h/cpp were essentially constant, in that
they were set in one place and never changed, but they were desfined as
regular ints, externed in Settings.h and defined in Settings.cpp.  As
such, the compiler is forced to allocate them in memory and every access
to them results in a memory access rather than using an immediate
constant value in the instruction.

This patch changes these values to static const int, a compile-time
constant and allows the compiler to compile them out.