Rework WidgetTooltip & LootDef
WidgetTooltip & TooltipData didn't have any real meaning or use apart
from each other so I've merged them.  Tooltip lines are now a vector of
strings and it has a mechanism to track when it's been changed, so it
knows when it's surface needs to be re-rendered.  Unfortunately, it
isn't used well yet, so it's still being re-rasterizted on most frames.

LootDef now manages a lot of it's own functionality and LootManager only
is only used for functionality that a LootDef object can't perform on
its own.

Made appropriate changes in other classes to support new WidgetTooltip
mechanism.  Managed to factor out a lot of code from MenuCharacter by
using some loops to populate tooltips for proficiencies.

Converted a numbef of global const ints to enums, as it fits much
better.

Stupid "trim trailing spaces" caused a lot of changes, even though I
tried to remove them all before, sorry :(