ANIMATED_TILES option {dis,en}ables background layer animation
If switched off, the background is not rendered all the time,
but only after movement of the player or if an event
happened.

This option cannot be set from within the config menu yet, but allows
the speedup in the backend to be easily used by setting
the ANIMATED_TILES variable in the config file.

The default is true, so no changes in comparison to the current
rendering are introduced.

The Tileset::logic function, which is responsible to calculate the
new frames in animations is also disabled on when disabling
the ANIMATED_TILES option.

Settings: correct string for animated tiles.