MapRenderer: micro optimize code.
* Skip all unneeded loop iterations at once.
  The loop iterations are skipped if you are at the border of the map.

* 3 out of 4 conditions were pulled out of the inner loop so it speeds up
  general rendering as well.

I was not able to measure the speedup. (Too much variance, top yields
a cpu usage from 28-30 now, mostly on 29 percent.
Before it was between 28-32, also most of the time at 29 percent.
So the variance of cpu usage in top is so high, I wouldn't really
claim I could *measure* the speedup.)